Changed validateFeature fn to accept a feature instead of a WKT feature representation and moved featureToWKT fn to map_functions file
